refactor: 회원가입 LazyColumn 스크롤 설정
Related to: #8
